Move rollcall to a library
