+++ src/dsbase/core/rollcall.h
@@ -3,6 +3,10 @@
  *
  *  Created on: May 9, 2018
  *      Author: djdup
+ *
+ * This manages the bulk sending of rollcall packets. It ensures that the
+ * packets go through even if the CAN bus is busy when the rollcall command is
+ * received.
  */
 
 #ifndef ROLLCALL_H_
@@ -16,7 +20,7 @@ typedef void (* rollcall_fn)(CANPacket *out);
 
 /**
  * Initialize rollcall with an array of functions. Each function populates a
- * CAN packet to send.
+ * CAN packet to send (but does not send it).
  */
 void rollcallInit(const rollcall_fn *functions, uint8_t numFunctions);
